tinyx86: Define addr_t as uint32_t
We are using 32 bit addresses. size_t is unreliable and is most likely
64-bit on most host platforms.

+++ include/tinyx86.h
@@ -8,4 +8,4 @@
 #define TINYX86_VERSION_MINOR 0
 #define TINYX86_VERSION_PATCH 1
 
-typedef size_t addr_t;
+typedef uint32_t addr_t;
